add database file, get userID
