+++ main.go
@@ -2,13 +2,7 @@ package main
 
 import (
 	"bot/disc"
-	"context"
-	"fmt"
-	"log"
 	"os"
-
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 func main() {
@@ -19,28 +13,4 @@ func main() {
 	}
 	path := dir + filename
 	disc.BotRun(path)
-
-}
-
-func database() {
-
-	// Set client options
-	clientOptions := options.Client().ApplyURI("mongodb://localhost:27017")
-
-	// Connect to MongoDB
-	client, err := mongo.Connect(context.TODO(), clientOptions)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Check the connection
-	err = client.Ping(context.TODO(), nil)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println("Connected to MongoDB!")
-
-	users := client.Database("waifu").Collection("users")
-	fmt.Println(users)
 }
